+++ seedfarmer/models/manifests/_module_manifest.py
@@ -71,9 +71,9 @@ class ModuleManifest(CamelModel):
 
     name: str
     path: str
-    bundle_md5: Optional[str]
-    manifest_md5: Optional[str]
-    deployspec_md5: Optional[str]
+    bundle_md5: Optional[str] = None
+    manifest_md5: Optional[str] = None
+    deployspec_md5: Optional[str] = None
     parameters: List[ModuleParameter] = []
     deploy_spec: Optional[DeploySpec] = None
     target_account: Optional[str] = None